src/services/weather-ai-outfit/pyproject.toml
[project]
name = "weather-ai-outfit"
version = "0.1.0"
description = "Weather AI Outfit service"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "openai>=1.0.0",
    "opentelemetry-distro>=0.59b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
]